pyproject.toml
[project]
name = "databricks-opentelemetry-issue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-monitor-opentelemetry>=1.6.12",
    "databricks-langchain>=0.6.0",
    "databricks-sdk>=0.60.0",
    "fastapi[cli,standard]>=0.116.1",
    "langchain>=0.3.27",
    "langgraph>=0.5.4",
    "mlflow>=3.1.4",
    "opentelemetry-instrumentation-fastapi>=0.52b1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
